Some updates for matter room AC
no custom capability is required as air purifier fan mode is in the main repository

moves most air purifier fan mode functionality to a sub driver
