+++ drivers/SmartThings/matter-thermostat/src/test/test_matter_thermo_featuremap.lua
@@ -27,7 +27,7 @@ local mock_device = test.mock_device.build_test_matter_device({
     {
       endpoint_id = 0,
       clusters = {
-        {cluster_id = clusters.Basic.ID, cluster_type = "SERVER"},
+        {cluster_id = clusters.BasicInformation.ID, cluster_type = "SERVER"},
       },
       device_types = {
         device_type_id = 0x0016, device_type_revision = 1, -- RootNode
@@ -61,7 +61,7 @@ local mock_device_simple = test.mock_device.build_test_matter_device({
     {
       endpoint_id = 0,
       clusters = {
-        {cluster_id = clusters.Basic.ID, cluster_type = "SERVER"},
+        {cluster_id = clusters.BasicInformation.ID, cluster_type = "SERVER"},
       },
       device_types = {
         device_type_id = 0x0016, device_type_revision = 1, -- RootNode
@@ -93,7 +93,7 @@ local mock_device_no_battery = test.mock_device.build_test_matter_device({
     {
       endpoint_id = 0,
       clusters = {
-        {cluster_id = clusters.Basic.ID, cluster_type = "SERVER"},
+        {cluster_id = clusters.BasicInformation.ID, cluster_type = "SERVER"},
       },
       device_types = {
         device_type_id = 0x0016, device_type_revision = 1, -- RootNode
